+++ src/dtu.blognet.Core.Command/CommandHandlerFactories/BlogCommandHandlerFactory.cs
@@ -6,17 +6,17 @@ namespace dtu.blognet.Core.Command.CommandHandlerFactories
 {
     public class BlogCommandHandlerFactory
     {
-      private readonly ApplicationDbContext _context;
-
-      public BlogCommandHandlerFactory(ApplicationDbContext context)
-      {
-        _context = context;
-      }
-
-      public ICommandHandler<AddBlogCommand, CommandResponse> Build(AddBlogCommand command)
-      {
-        return new AddBlogCommandHandler(_context, command);
-      }
-      
+        private readonly ApplicationDbContext _context;
+  
+        public BlogCommandHandlerFactory(ApplicationDbContext context)
+        {
+            _context = context;
+        }
+  
+        public ICommandHandler<AddBlogCommand, CommandResponse> Build(AddBlogCommand command)
+        {
+            return new AddBlogCommandHandler(_context, command);
+        }
+        
     }
 }
