+++ src/dtu.blognet.Core.Command/CommandHandlers/BlogCommandHandlers/AddBlogCommandHandler.cs
@@ -5,45 +5,45 @@ using dtu.blognet.Infrastructure.DataAccess;
 
 namespace dtu.blognet.Core.Command.CommandHandlers.BlogCommandHandlers
 {
-  public class AddBlogCommandHandler : BaseCommandHandler, ICommandHandler<AddBlogCommand, CommandResponse>
-  {
-    private readonly AddBlogCommand _command;
-
-    public AddBlogCommandHandler(ApplicationDbContext dbContext, AddBlogCommand command) : base(dbContext)
+    public class AddBlogCommandHandler : BaseCommandHandler, ICommandHandler<AddBlogCommand, CommandResponse>
     {
-      _command = command;
-    }
-
-
-    public CommandResponse Execute()
-    {
-        var response = new CommandResponse()
+        private readonly AddBlogCommand _command;
+    
+        public AddBlogCommandHandler(ApplicationDbContext dbContext, AddBlogCommand command) : base(dbContext)
         {
-          Success = false
-        };
-      
-        using (var transaction = _dbContext.Database.BeginTransaction())
-        {
-          // Should be automapper
-          var blog = new Blog
-          {
-              Title = _command.Model.Title
-          };
-          try
-          {
-              _dbContext.Blogs.Add(blog);
-              _dbContext.SaveChanges();
-              transaction.Commit();
-              response.Success = true;
-          }
-          catch (Exception e)
-          {
-            // LOG!!!
-            Console.WriteLine(e);
-            transaction.Rollback();
-          }
+            _command = command;
         }
-        return response;
+    
+    
+        public CommandResponse Execute()
+        {
+            var response = new CommandResponse()
+            {
+                Success = false
+            };
+          
+            using (var transaction = _dbContext.Database.BeginTransaction())
+            {
+                // Should be automapper
+                var blog = new Blog
+                {
+                    Title = _command.Model.Title
+                };
+                try
+                {
+                    _dbContext.Blogs.Add(blog);
+                    _dbContext.SaveChanges();
+                    transaction.Commit();
+                    response.Success = true;
+                }
+                catch (Exception e)
+                {
+                  // LOG!!!
+                  Console.WriteLine(e);
+                  transaction.Rollback();
+                }
+              }
+              return response;
+        }
     }
-  }
 }
