+++ src/dtu.blognet.Core.Command/ICommand.cs
@@ -1,7 +1,7 @@
 ﻿namespace dtu.blognet.Core.Command
 {
-  public interface ICommand<out TResult>
-  {
-    
-  }
+    public interface ICommand<out TResult>
+    {
+        
+    }
 }
