+++ src/dtu.blognet.Core.Command/InputModels/BlogInputModels/BlogInputModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace dtu.blognet.Core.Command.InputModels.BlogInputModels
 {
-  public class BlogInputModel
-  {
-    public string Title { get; set; }
-  }
+    public class BlogInputModel
+    {
+        public string Title { get; set; }
+    }
 }
